@@ -1,115 +1,77 @@
 {
-  "name": "vue",
-  "version": "1.0.0",
-  "description": "A Vue.js project",
-  "author": "",
+  "name": "open-doc-vue3",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "webpack-dev-server --inline --progress --config build/webpack.dev.conf.js --host 0.0.0.0",
-    "start": "npm run dev",
-    "unit": "jest --config test/unit/jest.conf.js --coverage",
-    "e2e": "node test/e2e/runner.js",
-    "test": "npm run unit && npm run e2e",
-    "lint": "eslint --ext .js,.vue src test/unit/specs test/e2e/specs",
-    "build": "node build/build.js"
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint",
+    "dev": "vue-cli-service serve"
   },
   "dependencies": {
-    "@tinymce/tinymce-vue": "^3.0.1",
-    "axios": "^0.17.1",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
-    "codemirror": "^5.65.12",
+    "@element-plus/icons-vue": "^2.3.1",
+    "axios": "^1.6.2",
+    "core-js": "^3.8.3",
     "default-passive-events": "^2.0.0",
-    "echarts": "^5.4.3",
-    "element-ui": "^2.15.13",
-    "exceljs": "^3.10.0",
+    "exceljs": "^4.4.0",
     "file-saver": "^2.0.5",
-    "js-md5": "^0.8.3",
-    "luckyexcel": "^1.0.1",
-    "resize-detector": "^0.3.0",
-    "scriptjs": "^2.5.8",
-    "shelljs": "^0.8.5",
-    "tinymce": "^5.1.0",
-    "url-search-params-polyfill": "^2.0.3",
-    "vue": "^2.7.6",
-    "vue-axios": "^2.0.2",
-    "vue-clipboard2": "^0.2.1",
-    "vue-codemirror": "^4.0.6",
-    "vue-hot-reload-api": "^2.3.4",
-    "vue-i18n": "^8.28.2",
-    "vue-print-nb": "^1.7.5",
-    "vue-router": "^3.0.1",
-    "vuex": "^3.1.0",
-    "whatwg-fetch": "^2.0.3"
+    "puppeteer": "^21.6.0",
+    "sass": "^1.69.5",
+    "sass-loader": "^13.3.2",
+    "scriptjs": "^2.5.9",
+    "vue": "^3.2.13",
+    "vue-axios": "^3.5.2",
+    "vue-class-component": "^8.0.0-0",
+    "vue-i18n": "^9.8.0",
+    "vue-loader": "^17.3.1",
+    "vue-router": "^4.2.5",
+    "vue3-print-nb": "^0.1.4",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "autoprefixer": "^7.1.2",
-    "axios": "^0.17.1",
-    "babel-core": "^6.26.3",
-    "babel-eslint": "^7.1.1",
-    "babel-helper-vue-jsx-merge-props": "^2.0.3",
-    "babel-jest": "^21.0.2",
-    "babel-loader": "^7.1.1",
-    "babel-plugin-dynamic-import-node": "^1.2.0",
-    "babel-plugin-syntax-jsx": "^6.18.0",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
-    "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-plugin-transform-vue-jsx": "^3.5.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-register": "^6.22.0",
-    "chalk": "^2.0.1",
-    "compression-webpack-plugin": "^1.1.1",
-    "copy-webpack-plugin": "^4.5.2",
-    "cross-spawn": "^5.0.1",
-    "css-loader": "^0.28.0",
-    "eslint": "^3.19.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-friendly-formatter": "^3.0.0",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-html": "^3.0.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.2.0",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-standard": "^3.0.1",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^1.1.4",
-    "friendly-errors-webpack-plugin": "^1.6.1",
-    "html-webpack-plugin": "^2.30.1",
-    "jest": "^21.2.0",
-    "jest-serializer-vue": "^0.3.0",
-    "nightwatch": "^0.9.12",
-    "node-notifier": "^5.1.2",
-    "optimize-css-assets-webpack-plugin": "^3.2.0",
-    "ora": "^1.2.0",
-    "portfinder": "^1.0.13",
-    "postcss-import": "^11.0.0",
-    "postcss-loader": "^2.0.8",
-    "rimraf": "^2.6.0",
-    "sass": "^1.52.3",
-    "sass-loader": "^7.3.1",
-    "scriptjs": "^2.5.8",
-    "selenium-server": "^3.0.1",
-    "semver": "^5.3.0",
-    "uglifyjs-webpack-plugin": "^1.1.1",
-    "url-loader": "^0.5.8",
-    "vue-axios": "^2.0.2",
-    "vue-jest": "^1.0.2",
-    "vue-loader": "^13.3.0",
-    "vue-style-loader": "^3.0.1",
-    "vue-template-compiler": "^2.5.2",
-    "webpack": "^3.6.0",
-    "webpack-bundle-analyzer": "^2.9.0",
-    "webpack-dev-server": "^2.9.1",
-    "webpack-merge": "^2.6.1"
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@codemirror/lang-javascript": "^0.20.0",
+    "@codemirror/theme-one-dark": "^0.20.0",
+    "@tinymce/tinymce-vue": "^5.1.1",
+    "@typescript-eslint/eslint-plugin": "^5.4.0",
+    "@typescript-eslint/parser": "^5.4.0",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "@vue/eslint-config-typescript": "^9.1.0",
+    "codemirror": "^5.65.12",
+    "compression-webpack-plugin": "^10.0.0",
+    "echarts": "^5.4.3",
+    "element-plus": "^2.4.3",
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.0.3",
+    "js-md5": "^0.8.3",
+    "luckyexcel": "^1.0.1",
+    "prerender-spa-plugin-next": "^4.2.3",
+    "resize-detector": "^0.3.0",
+    "tinymce": "^5.10.9",
+    "vue-clipboard3": "^2.0.0",
+    "vue-codemirror": "^5.0.1"
   },
-  "engines": {
-    "node": ">= 4.0.0",
-    "npm": ">= 3.0.0"
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended",
+      "@vue/typescript"
+    ],
+    "parserOptions": {
+      "parser": "@typescript-eslint/parser"
+    },
+    "rules": {}
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not ie <= 8"
+    "not ie 11"
   ]
 }
