@@ -1,14 +1,12 @@
-// The Vue build version to load with the `import` command
-// (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue'
-import App from './App'
-import router from './router'
-import ElementUI from 'element-ui'
-import 'element-ui/lib/theme-chalk/index.css'
-import '../static/icon/iconfont.css';
-import "../static/css/custom-element.css";
-//在min.js中引入插件
-import 'default-passive-events'; // 解决事件报错问题https://blog.csdn.net/tonetwo/article/details/129384680
+import { createApp } from 'vue'
+import App from './App.vue'
+
+import router from '@/router'
+import store from '@/store'
+import ElementPlus from 'element-plus'
+import 'element-plus/theme-chalk/index.css'
+import '../public/static/icon/iconfont.css';
+import '../public/static/css/custom-element.css';
 
 import Header from '@/components/common/Header'
 import Footer from '@/components/common/Footer'
@@ -16,28 +14,29 @@ import util from '@/js/util.js'
 import axios from '@/js/http'
 
 import i18n from '@/lang'
-import 'url-search-params-polyfill'
-import "babel-polyfill";
-import VueClipboard from 'vue-clipboard2';
+// import 'url-search-params-polyfill'
+// import "babel-polyfill";
 
-Vue.use(util);
-Vue.config.productionTip = false
-Vue.component('Header', Header);
-Vue.component('Footer', Footer);
-Vue.use(ElementUI)
-Vue.use(VueClipboard)
+import 'default-passive-events' // 解决事件报错问题https://blog.csdn.net/tonetwo/article/details/129384680
+import '@/js/browserPatch.js'
 
-// 将axios挂载到prototype上，在组件中可以直接使用this.axios访问
-Vue.prototype.axios = axios;
+// 如果您正在使用CDN引入，请删除下面一行。
+import * as ElementPlusIconsVue from '@element-plus/icons-vue'
+import DirectiveExtensions from '@/directive' //自定义指令集（比如：vue-clipboard3）
 
 
-/* eslint-disable no-new */
-
-new Vue({
-  el: '#app',
-  router,
-  i18n,
-  template: '<App/>',
-  components: { App }
-})
-
+const app = createApp(App);
+app.config.globalProperties.$axios = axios;// 将axios挂载到prototype上，在组件中可以直接使用this.axios访问
+app.component('Header', Header)
+    .component('Footer', Footer)
+    .use(util)
+    .use(ElementPlus)
+    .use(store)
+    .use(i18n)
+    .use(router)
+    .use(DirectiveExtensions)
+    .mount('#app');
+// 引入icon
+for (const [key, component] of Object.entries(ElementPlusIconsVue)) {
+    app.component(key, component)
+}
