@@ -199,7 +199,7 @@ class ItemModel extends BaseModel {
 
     //获取菜单结构
     public function getMemu($item_id){
-            $page_field = "page_id,author_uid,cat_id,page_title,addtime,s_number";
+            $page_field = "page_id,author_uid,cat_id,page_title,addtime,s_number,page_use";
             $catalog_field = '*';
             $data = $this->getContent($item_id , $page_field , $catalog_field) ;
             return $data ;
