@@ -12,6 +12,6 @@ export default {
 
 <style>
 
- @import "../static/css/common.css";
+/*@import "../static/css/common.css";*/
 
 </style>
