@@ -8,15 +8,15 @@ export default {
     section_title1: 'OpenDoc',
     section_description1: 'Pursuing the ideal of IT and becoming the best document.',
     section_title2: 'API Document',
-    section_description2: 'OpenDoc can compile exquisite API documents <br>in a very fast and convenient way',
+    section_description2: 'OpenDoc can compile exquisite API documents in a very fast and convenient way',
     section_title3: 'Data Dictionary',
-    section_description3: 'A good Data Dictionary can easily exhibit database structure to other people<br>OpenDoc can compile exquisite Data Dictionary',
+    section_description3: 'A good Data Dictionary can easily exhibit database structure to other people OpenDoc can compile exquisite Data Dictionary',
     section_title4: 'Explanation Document',
     section_description4: 'You can absolutely use OpenDoc to compile the explanation documents for some tools',
     section_title5: 'Team Work',
     section_description5: 'Your team will work with OpenDoc together very well ',
     section_title6: 'Open Source',
-    section_description6: 'OpenDoc is a free, open source tool that <br>you can deploy it to your own server',
+    section_description6: 'OpenDoc is a free, open source tool that  you can deploy it to your own server',
     section_title7: 'Hosted online',
     section_description7: 'open-doc.docker-sky.cn provide security and stability of the document hosting service',
     section_title8: 'Try it now',
@@ -135,10 +135,10 @@ export default {
     archive_tips2: 'Explanation: After archiving the project, it will no longer be able to add or modify content, and the archiving status cannot be cancelled. If you want to modify the content again, you can copy this project and modify it based on the new one. The method of copying a project is to choose to copy from an existing project when creating it.',
     success_jump: 'The operation is successful! Skipping...',
     reset_token: 'Reset Token',
-    open_api_tips1: 'OpenDoc is an API for open document editing, making it easier for users to manipulate document data<br/> By utilizing open APIs, automation functions have been supported as follows :',
-    open_api_tips2: 'If you want to automate the generation of API documents, you can refer to 【<a target="_bank" href="'+DocConfig.homeUrl+'/item-show/36?page_id=3954">Api Doc</a>】',
-    open_api_tips3: 'If you want to automate the generation of a data dictionary, you can refer to 【<a target="_bank" href="'+DocConfig.homeUrl+'/item-show/36?page_id=3955">Data Dictionary</a>】',
-    open_api_tips4: 'If you are more free to generate the format you need, you can refer to 【<a target="_bank" href="'+DocConfig.homeUrl+'/item-show/36?page_id=3956">Open Api</a>】',
+    open_api_tips1: 'OpenDoc is an API for open document editing, making it easier for users to manipulate document data By utilizing open APIs, automation functions have been supported as follows :',
+    open_api_api:'Api Doc',
+    open_api_data:'Data Dictionary',
+    open_api_custom:'Open Api',
     //item/show
     item: 'Item',
     share: 'Share',
@@ -151,6 +151,9 @@ export default {
     new_page: 'New Page',
     edit_page: 'Edit Page',
     new_catalog: 'New Folder',
+    place_input_tip:'Please input your ',
+    empty_input:'The input content cannot be empty.',
+    limit_length_input:'The length of the input content cannot be less than 3.',
     share_address_to_your_friends: 'You can share the address to your friends',
     copy_interface_to_new: 'Copy this page to a new page',
     copy: 'Copy',
@@ -178,7 +181,8 @@ export default {
     default_new_doc:'Default New Document',
     close_drag:'Close Drag',
     open_drag:'Open Drag',
-    project_share_and_page_share_diff_tip:'<p>Project page:</p><p>Shared users can access other pages under the same project with their own permissions through the page URL</p><p>Single page:</p><p>Shared users can only see the current page</p>',
+    project_share_tip:'Shared users can access other pages under the same project with their own permissions through the page URL.',
+    page_share_tip:'Shared users can only see the current page.',
 
     //page
     title: 'Title',
@@ -188,7 +192,7 @@ export default {
     s_number: 'Order',
     s_number_explain: '(Optional) the default is 99, the smaller the number, the earlier',
     optional: 'Optional',
-    history_version: 'History version',
+    history_version: 'Versions',
     save: 'Save',
     cancel: 'Cancel',
     insert_job_plan_template: 'Insert JobPlan Template',
@@ -200,6 +204,11 @@ export default {
     doc_tools:'Tools',
     quick_tools:'Quick tools',
     editor:'Editor',
+    clean_json:'Clean',
+    format_json:'Format Json',
+    insert_json:'Insert Json',
+    insert_table:'Insert Table',
+    insert_parameter_json:'Insert Parameter Table',
     word_editor:'Word editor',
     markdowm_editor:'Markdowm editor',
     excel_editor:'Excel editor',
@@ -237,6 +246,7 @@ export default {
     update_by_who: 'Update by ',
     operation: 'Operation',
     recover_to_this_version: 'Recover',
+    empty_version_history:'There is currently no version history information available',
     close: 'Close',
     finish: 'Finish',
     params: 'Params',
@@ -244,6 +254,7 @@ export default {
     result: 'Result',
     save_to_templ: 'Save to template',
     more_templ: 'My Templpate',
+    default_templ:'Default Templpate',
     saved_templ_list: 'Template List',
     page_comments: 'Page Notes',
     cur_page_content: 'Current Newest Version',
@@ -305,6 +316,7 @@ export default {
 
     copy_link: 'Copy',
     copy_success: 'Copy Success',
+    copy_error: 'Copy Error',
     doc_edit_area: 'Doc edit area',
     empty_data: 'Empty data',
     doc_empty: 'Please click on the editing mode at the top left of the page or the editing button on the right to edit it.',
