@@ -1,20 +1,17 @@
-import Vue from 'vue'
-import VueI18n from 'vue-i18n'
-Vue.use(VueI18n)
+import { createI18n } from 'vue-i18n';
 
-import zh from './zh'
-import en from './en'
-
+import ZH from './zh.js';
+import EN from './en.js';
 const messages = {
-  'zh': zh,
-  'en': en,
-}
+  zh: { ...ZH  },
+  en: { ...EN  },
+};
 
-const i18n = new VueI18n({
+const i18n = createI18n({
   // 初始语言找本地储存的lang，默认为浏览器语言
-  locale:  localStorage.lang  || 'zh',
-  fallbackLocale: 'zh-CN',
-  messages,
-})
-
-export default i18n
+  legacy: false,
+  globalInjection: true,
+  locale: localStorage.lang  || 'zh',
+  messages
+});
+export default i18n;
