@@ -8,27 +8,27 @@ export default {
     section_title1: 'OpenDoc',
     section_description1: '积叶成书，做最好用的IT文档.',
     section_title2: 'API文档',
-    section_description2: ' APP、web前端与服务器<br>常用API来进行交互',
+    section_description2: ' APP、web前端与服务器常用API来进行交互',
     section_title3: '数据字典',
-    section_description3: '好的数据字典可以很方便地<br>向别人描述你的数据库结构',
+    section_description3: '好的数据字典可以很方便地向别人描述你的数据库结构',
     section_title4: '说明文档',
-    section_description4: 'OpenDoc可以编写一些工具的说明书<br>或编写一些技术规范说明文档以供团队查阅',
+    section_description4: 'OpenDoc可以编写一些工具的说明书或编写一些技术规范说明文档以供团队查阅',
     section_title5: '团队协作',
-    section_description5: '完善的团队权限管理机制助力团队<br>良好高效地协同编写文档',
+    section_description5: '完善的团队权限管理机制助力团队良好高效地协同编写文档',
     section_title6: '免费开源',
-    section_description6: 'OpenDoc提供免费开源的版本<br>您可以选择将OpenDoc部署到您的服务器',
+    section_description6: 'OpenDoc提供免费开源的版本 您可以选择将OpenDoc部署到您的服务器',
     section_title7: '在线托管',
-    section_description7: DocConfig.homeUrl +' 提供安全稳定的在线文档托管服务<br>您可以放心地选择托管你的文档数据在云端',
+    section_description7: DocConfig.homeUrl +' 提供安全稳定的在线文档托管服务 您可以放心地选择托管你的文档数据在云端',
     section_title8: '立即体验',
     section_description8: '超过6000+互联网团队正在使用OpenDoc',
     section_title9: '共享文档',
     section_description9: '支持多人协作编辑在线excel文档',
     section_title10: '产品文档',
-    section_description10: 'OpenDoc集成了markdown编辑器<br>高效支持产品从概念化进入到图纸化',
+    section_description10: 'OpenDoc集成了markdown编辑器 高效支持产品从概念化进入到图纸化',
     section_title11: '办公日记',
-    section_description11  : '好记性不如烂笔头<br>OpenDoc是工作办公的最佳写字板',
+    section_description11  : '好记性不如烂笔头 OpenDoc是工作办公的最佳写字板',
     section_title12: '工作总结',
-    section_description12: '做的好不如总结的好<br>总结得好离不开OpenDoc的完美支持',
+    section_description12: '做的好不如总结的好 总结得好离不开OpenDoc的完美支持',
 
     //user
     login:'登录',
@@ -140,10 +140,10 @@ export default {
     archive_tips2: '说明： 归档项目后，项目将无法再新增和修改内容，并且无法取消归档状态。 如想再次修改内容，可复制本项目，在新的项目基础上修改。复制项目的方法是，在创建项目的时候，选择从已有项目里复制。',
     success_jump: '操作成功！正在跳转...',
     reset_token: '重新生成api_token',
-    open_api_tips1: 'OpenDoc开放文档编辑的API，供使用者更加方便地操作文档数据。<br>利用开放API，已支持自动化功能如下：',
-    open_api_tips2: '如果您想自动化生成API文档，则可参考 【<a target="_bank" href="'+DocConfig.homeUrl+'/item-show/36?page_id=3954">API文档</a>】',
-    open_api_tips3: '如果您想自动化生成数据字典，则可参考 【<a target="_bank" href="'+DocConfig.homeUrl+'/item-show/36?page_id=3955">数据字典</a>】',
-    open_api_tips4: '如果您更自由地生成自己所需要的格式，则可参考 【<a target="_bank" href="'+DocConfig.homeUrl+'/item-show/36?page_id=3956">开放API</a>】',
+    open_api_tips1: 'OpenDoc开放文档编辑的API，供使用者更加方便地操作文档数据。 利用开放API，已支持自动化功能如下：',
+    open_api_api:'API文档',
+    open_api_data:'数据字典',
+    open_api_custom:'开放API',
 
     //item/show
     item:'项目',
@@ -157,12 +157,15 @@ export default {
     new_page:'文档',
     edit_page:'编辑文档',
     new_catalog:'文件夹',
+    place_input_tip:'请输入您的 ',
+    empty_input:'输入内容不能为空.',
+    limit_length_input:'输入内容长度不能小于3.',
     share_address_to_your_friends:'分享该接口地址给你的好友',
     share:'分享',
     copy_interface_to_new:'复制该页面到新页面',
     copy:'复制',
     edit:'编辑',
-    detail:'详情',
+    detail:'历史操作',
     delete_interface:'删除页面',
     comfirm_delete:'确认删除吗？',
     delete:'删除',
@@ -185,7 +188,8 @@ export default {
     default_new_doc:'默认新文档',
     close_drag:'退出拖拽',
     open_drag:'拖拽模式',
-    project_share_and_page_share_diff_tip:'<p>项目页面：</p><p>被分享的用户可以通过页面url访问同项目下自身有权限的其他页面.</p><p>单页面：</p><p>被分享用户仅可以看到当前页面.</p>',
+    project_share_tip:'被分享的用户可以通过页面url访问同项目下自身有权限的其他页面.',
+    page_share_tip:'被分享用户仅可以看到当前页面.',
 
     //page
     title:'标题',
@@ -207,6 +211,11 @@ export default {
     json_tools:'JSON工具',
     quick_tools:'便捷工具',
     editor:'编辑器',
+    clean_json:'清空内容',
+    format_json:'格式化校验',
+    insert_json:'插入Json',
+    insert_table:'插入表格',
+    insert_parameter_json:'插入参数表格',
     word_editor:'word文档',
     markdowm_editor:'markdowm文档',
     excel_editor:'excel表格',
@@ -246,6 +255,7 @@ export default {
     update_by_who:'修改人',
     operation:'操作',
     recover_to_this_version:'恢复到此版本',
+    empty_version_history:'暂无版本历史信息',
     close:'关闭',
     finish:'完成',
     params:'参数',
@@ -253,6 +263,7 @@ export default {
     result:'返回结果',
     save_to_templ:'另存为模板',
     more_templ:'我的模板',
+    default_templ:'默认模板',
     saved_templ_list:'保存的模板列表',
     page_comments:'页面注释',
     cur_page_content:'当前最新版本',
@@ -315,6 +326,7 @@ export default {
 
     copy_link: '复制',
     copy_success: '复制成功',
+    copy_error: '复制失败',
     doc_edit_area: '文档编辑区',
     empty_data: '暂无数据',
     doc_empty: '请点击页面左上方的【编辑模式】 或 右侧的【编辑】按钮进行编辑.',
